refactor(performance): extract business logic to PerformanceReportService
- Extract calculatePerformanceScore() and generateOverallRecommendations() to service layer
- Move CacheMetrics and DatabasePerformanceMetrics type definitions to service
- Reduce route file from 175 to 47 lines (73% reduction)
- Follow blueprint.md:208-209 Service Layer principles
- Improve testability and reusability of performance analysis logic
- Zero breaking changes - same API contract maintained
